updated bitcoin and lightningd deployment scripts

+++ multipass/lxd/bcm/bitcoin/down_bitcoin.sh
@@ -26,32 +26,32 @@ if [[ $(lxc exec manager1 -- docker stack ls | grep bitcoind) ]]; then
     lxc exec manager1 -- docker stack rm bitcoind
 fi
 
+if [[ $(lxc exec bitcoin -- docker info | grep "Swarm: active") ]]; then
+    echo "Removing docker daemon in lxd host 'bitcoin' from the swarm."
+    lxc exec bitcoin -- docker swarm leave
+fi
 
 if [[ $(lxc list | grep bitcoin) ]]; then
     echo "Destroying lxd container 'bitcoin'."
     lxc delete --force bitcoin
-else
-    echo "LXC container 'bitcoin' not found. Skipping."
 fi
 
 if [[ $(lxc profile list | grep bitcoinprofile) ]]; then
     echo "Destroying lxd profile 'bitcoinprofile'."
    lxc profile delete bitcoinprofile
-else
-    echo "LXC profile 'bitcoinprofile' not found. Skipping."
 fi
 
 
 if [[ $(lxc network list | grep lxdbrBitcoin) ]]; then
     echo "Destroying lxd network 'lxdbrBitcoin'."
    lxc network delete lxdbrBitcoin
-else
-    echo "LXC network 'lxdbrBitcoin' not found. Skipping."
 fi
 
-if [[ $(lxc storage list | grep "bitcoin-dockervol") ]]; then
-    echo "Destroying lxd storage pool 'bitcoin-dockervol'."
-   lxc storage delete bitcoin-dockervol
-else
-    echo "LXC storage pool 'bitcoin-dockervol' not found. Skipping."
+
+# if the user has instructed us to delete the dockervol backing.
+if [[ $BCM_BITCOIN_DELETE_DOCKERVOL = "true" ]]; then
+    if [[ $(lxc storage list | grep "bitcoin-dockervol") ]]; then
+        echo "Destroying lxd storage pool 'bitcoin-dockervol'."
+        lxc storage delete bitcoin-dockervol
+    fi
 fi
