+++ multipass/lxd/bcs/down_lxd.sh
@@ -25,8 +25,6 @@ lxc list >>/dev/null
 if [[ $(lxc list | grep cachestack) ]]; then
     echo "Deleting lxd container 'cachestack'."
     lxc delete --force cachestack >/dev/null
-else
-    echo "Skipping deletion of lxd container 'cachestack'."
 fi
 
 
@@ -34,8 +32,6 @@ fi
 if [[ $(lxc profile list | grep cachestackprofile) ]]; then
     echo "Deleting lxd profile 'cachestackprofile'."
     lxc profile delete cachestackprofile >/dev/null
-else
-    echo "Skipping deletion of lxd profile 'cachestackprofile'."
 fi
 
 
@@ -43,8 +39,6 @@ fi
 if [[ $(lxc network list | grep lxdbrCacheStack) ]]; then
     echo "Deleting lxd network 'lxdbrCacheStack'."
     lxc network delete lxdbrCacheStack
-else
-    echo "Skipping deletion of lxd network lxdbrCacheStack."
 fi
 
 
@@ -52,16 +46,21 @@ fi
 if [[ $(lxc network list | grep lxdBCSMgrnet) ]]; then
     echo "Deleting lxd network 'lxdBCSMgrnet'."
     lxc network delete lxdBCSMgrnet
-else
-    echo "Skipping deletion of lxd network lxdBCSMgrnet."
 fi
 
-# delete lxd storage cachestack-dockervol 
-if [[ $(lxc storage list | grep "cachestack-dockervol") ]]; then
-    echo "Deleting lxd storage pool 'cachestack-dockervol'."
-    lxc storage delete cachestack-dockervol
-else
-    echo "Skipping deletion of lxd storage pool cachestack-dockervol."
+# delete lxd network lxdBrNowhere 
+if [[ $(lxc network list | grep lxdBrNowhere) ]]; then
+    echo "Deleting lxd network 'lxdBrNowhere'."
+    lxc network delete lxdBrNowhere
+fi
+
+# if the user has instructed us to delete the dockervol backing.
+if [[ $BCS_DELETE_DOCKERVOL = "true" ]]; then
+    # delete lxd storage cachestack-dockervol 
+    if [[ $(lxc storage list | grep "cachestack-dockervol") ]]; then
+        echo "Deleting lxd storage pool 'cachestack-dockervol'."
+        lxc storage delete cachestack-dockervol
+    fi
 fi
 
 # delete the host template if configured
